+++ aitune/torch/backend/torch_inductor_jit_backend.py
@@ -63,6 +63,11 @@ class TorchInductorJitBackendConfig(BackendConfig):
         autocast_enabled (bool): If True, enable autocast.
 
         autocast_dtype (torch.dtype): The dtype to use for autocast.
+
+        Note: inference is done with torch.no_grad() context. The torch.inference_mode() context must not be used
+        as it would require outputs from a model to be used with same inference mode - this would be confusing to a user
+        and required code changes from the user.
+
     """
 
     fullgraph: bool = False
@@ -114,7 +119,7 @@ class TorchInductorJitBackend(Backend):
         # build variables
         self._compiled_module = None
         self._orig_module = None
-        self._output_dtype = None
+        self._required_casting_dtype = None
         self._data = None
         self._compile_dynamic = self._config.dynamic
 
@@ -126,23 +131,35 @@ class TorchInductorJitBackend(Backend):
         """Returns the description of the backend."""
         return f"{self.__class__.__name__}({self._config.describe()})"
 
-    @staticmethod
-    def _get_dtype(module: nn.Module, data: list[Sample]) -> torch.dtype:
-        """Get the output dtype of the module by running a sample inference.
+    def _get_required_casting_dtype(self, module: nn.Module, data: list[Sample]) -> torch.dtype | None:
+        """Get the required casting dtype of the module by running a sample inference with and without autocast.
+
+        If the dtype of the output is different with and without autocast, return the dtype of the output without autocast.
+        Otherwise, return None.
 
         Args:
             module (nn.Module): The module to get the dtype from.
             data (list[Sample]): List of sample inputs to run through the module.
 
         Returns:
-            torch.dtype: The dtype of the module's output tensor. Returns None if output is not a tensor.
+            torch.dtype: The required casting dtype. Returns None if no casting is required.
         """
-        args, kwargs = data[0]
-        res = module(*deepcopy(args), **deepcopy(kwargs))
-        if isinstance(res, torch.Tensor):
-            return res.dtype
-        else:
-            return None
+        with torch.no_grad():
+            with torch.autocast(
+                device_type=str(self._device.type),
+                dtype=self._config.autocast_dtype,
+                enabled=True,
+            ):
+                args, kwargs = deepcopy(data[0])
+                autocast_result = module(*args, **kwargs)
+
+            if isinstance(autocast_result, torch.Tensor):
+                args, kwargs = deepcopy(data[0])
+                orig_result = module(*args, **kwargs)
+                if orig_result.dtype != autocast_result.dtype:
+                    return orig_result.dtype
+
+        return None
 
     def _build(self, module: nn.Module, graph_spec: GraphSpec, data: list[Sample], cache_dir: Path) -> Backend:
         """Builds the model with torch.compile."""
@@ -150,10 +167,12 @@ class TorchInductorJitBackend(Backend):
         self._save_config(cache_dir)
 
         module.to(self._device)
-        self._output_dtype = self._get_dtype(module, data)
         self._orig_module = module
         self._data = data
+        if self._config.autocast_enabled:
+            self._required_casting_dtype = self._get_required_casting_dtype(module, data)
         self._compile()
+        self._activate()
         return self
 
     def _compile(self):
@@ -183,8 +202,10 @@ class TorchInductorJitBackend(Backend):
         """Activates backend."""
         if self._compiled_module is None:  # TBD pb: after introducing module states this should be changed
             self._compile()
+        if self._config.autocast_enabled:
+            self._infer = self._infer_with_autocast
 
-    def _infer(self, *args: Any, **kwargs: Any) -> Any:
+    def _infer_with_autocast(self, *args: Any, **kwargs: Any) -> Any:
         """Runs inference with the given arguments.
 
         Args:
@@ -194,20 +215,32 @@ class TorchInductorJitBackend(Backend):
         Returns:
             Any: The result of the inference.
         """
-        with (
-            torch.autocast(
-                device_type=str(self._device),
+        with torch.no_grad():
+            with torch.autocast(
+                device_type=str(self._device.type),
                 dtype=self._config.autocast_dtype,
-                enabled=self._config.autocast_enabled,
-            ),
-            torch.no_grad(),
-        ):
-            res = self._compiled_module(*args, **kwargs)
-            if isinstance(res, torch.Tensor) and res.dtype != self._output_dtype and self._output_dtype is not None:
-                # autocast changed the dtype of the output, converting back to the original dtype
-                return res.to(self._output_dtype)
-            else:
-                return res
+                enabled=True,
+            ):
+                res = self._compiled_module(*args, **kwargs)
+                if self._required_casting_dtype is not None:
+                    # autocast changed the dtype of the output, converting back to the original dtype
+                    return res.to(self._required_casting_dtype)
+        return res
+
+    def _infer(self, *args: Any, **kwargs: Any) -> Any:
+        """Runs inference with the given arguments. Does not use autocast.
+
+        It can be replaced at runtime by _infer_with_autocast.
+
+        Args:
+            *args: inference arguments
+            **kwargs: inference keyword arguments
+
+        Returns:
+            Any: The result of the inference.
+        """
+        with torch.no_grad():
+            return self._compiled_module(*args, **kwargs)
 
     def _deactivate(self):
         """Deactivates backend."""
@@ -233,7 +266,7 @@ class TorchInductorJitBackend(Backend):
         return {
             self.STATE_TYPE: self.__class__.__name__,
             self.STATE_CONFIG: self._config.to_dict(),
-            self.STATE_OUTPUT_DTYPE: self._output_dtype,
+            self.STATE_OUTPUT_DTYPE: self._required_casting_dtype,
             self.STATE_DATA: self._data,
             self.STATE_ORIG_MODULE: self._orig_module.state_dict(),
             self.STATE_DEVICE: self._device,
@@ -252,7 +285,7 @@ class TorchInductorJitBackend(Backend):
         config = TorchInductorJitBackendConfig.from_dict(state_dict[cls.STATE_CONFIG])
 
         backend = cls(config=config)
-        backend._output_dtype = state_dict[cls.STATE_OUTPUT_DTYPE]
+        backend._required_casting_dtype = state_dict[cls.STATE_OUTPUT_DTYPE]
         backend._data = state_dict[cls.STATE_DATA]
         backend._device = state_dict[cls.STATE_DEVICE]
         backend._compile_dynamic = state_dict.get(cls.STATE_COMPILE_DYNAMIC, config.dynamic)